+++ qml/media/MediaControls.qml
@@ -26,6 +26,7 @@ Rectangle {
 
     signal playPauseActivated()
     signal loadActivated()
+    signal saveAsActivated()
 
     function showControls() {
         controlHideTimer.restart();
@@ -91,25 +92,46 @@ Rectangle {
             spacing: 4
             Layout.fillWidth: true
 
-            // Cache/Play/pause button
+            // Cache button
+            ImageButton {
+                id: downloadImage
+
+                Layout.alignment: Qt.AlignLeft
+                buttonTextColor: GlobalObject.colors.text
+                Layout.preferredHeight: 20
+                Layout.preferredWidth: 20
+                image: {
+                    if (control.mediaLoaded) {
+                        enabled = true
+                        return ":/images/save.svg";
+                    } else {
+                        return ":/images/download.svg";
+                    }
+                }
+                onClicked: {
+                    if(control.mediaLoaded){
+                        control.saveAsActivated()
+                    } else {
+                        enabled = false
+                        control.loadActivated()
+                    }
+                }
+            }
+            // Play/pause button
             ImageButton {
                 id: playbackStateImage
-
+                enabled: control.mediaLoaded
                 Layout.alignment: Qt.AlignLeft
                 buttonTextColor: GlobalObject.colors.text
                 Layout.preferredHeight: 24
                 Layout.preferredWidth: 24
                 image: {
-                    if (control.mediaLoaded) {
-                        if (control.mediaState == MediaPlayer.PlayingState)
-                            return ":/images/pause-symbol.svg";
-                        else
-                            return ":/images/play-sign.svg";
-                    } else {
-                        return ":/images/download.svg";
-                    }
+                    if (control.mediaState == MediaPlayer.PlayingState)
+                        return ":/images/pause-symbol.svg";
+                    else
+                        return ":/images/play-sign.svg";
                 }
-                onClicked: control.mediaLoaded ? control.playPauseActivated() : control.loadActivated()
+                onClicked: control.playPauseActivated()
             }
 
             ImageButton {